Rename td-selector module to members to be more generic
