+++ src/attendance-input/index.jsx
@@ -5,7 +5,7 @@ import { Tab, Tabs, TabList, TabPanel } from "react-tabs";
 import "react-tabs/style/react-tabs.css";
 
 import dailMembers from "../../data/dail/32/members.json";
-import { TDSelector } from "../td-selector/TDSelector.jsx";
+import { TDSelector } from "../members/TDSelector.jsx";
 import { AttendanceInputText } from "./AttendanceInputText.jsx";
 import { AttendanceInputCalendar } from "./AttendanceInputCalendar.jsx";
